Update header and navigation
