+++ app/layout.tsx
@@ -1,3 +1,4 @@
+import { Header } from '@/components/landing/header'
 import { inter, roboto_mono } from '@/lib/fonts'
 import Providers from '@/lib/providers'
 import '@/styles/globals.css'
@@ -16,6 +17,7 @@ export default function RootLayout(props: Props) {
 	return (
 		<html lang="en" className={`${inter.variable} ${roboto_mono.variable} h-full`} suppressHydrationWarning>
 			<body className="h-full bg-background text-foreground">
+				<Header/>
 				<Providers>{props.children}</Providers>
 			</body>
 		</html>
